board/raspberrypi: include .config dfs files.
This change pulls dts filenames from the .config file instead of
specifying the required dts files in a board config file.
This reduces duplication and makes it possible to specify different
dts files in the config file to have it appear in the resultant
image. You would want to do so if you were using upstream kernel
sources.